requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.1.3
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
JPype1==1.4.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
